% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glod.R
\name{glod}
\alias{glod}
\title{Plot the glod, a joke ( Geographer's Line of Death)}
\usage{
glod(x = "+proj=laea", plot = TRUE, n = 361)
}
\arguments{
\item{x}{a projection to use, a string understood by PROJ itself}

\item{plot}{logical, will create a plot by default}

\item{n}{number of points to plot on the line}
}
\value{
the line of points, invisibly if `plot = TRUE``
}
\description{
Creates a map and plots the line where lat == lon, because no other places are as confusing.
}
\examples{
glod()
glod("+proj=laea +lon_0=147 +lat_0=-42")
}
