% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glod-package.R
\docType{package}
\name{glod-package}
\alias{glod-package}
\alias{_PACKAGE}
\title{glod: The Geographer's Line of Death}
\description{
Plot the geograper's line of death, with options.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mdsumner/glod}
  \item Report bugs at \url{https://github.com/mdsumner/glod/issues}
}

}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
